requirements.txt
pandas
pytest
streamlit
python-dotenv
google-genai
